use custom weights instead of duration.

+++ prepare.sh
@@ -5,41 +5,54 @@ echo "munichways routing - setup"
 echo "=========================="
 echo
 
+SKIP_ANNOTATION=false
+if [ "$1" = "--skip-annotation" ]; then
+  SKIP_ANNOTATION=true
+fi
+
 # cleanup last run
-echo "cleaning up ..."
-mkdir -p ./out
-mkdir -p ./geo
-rm ./out/out.pbf 2>/dev/null
-rm ./out/out.geojson 2>/dev/null
-echo "done."
-echo
+if [ "$SKIP_ANNOTATION" = false ]; then
+  echo "cleaning up ..."
+  mkdir -p ./out
+  mkdir -p ./geo
+  rm ./out/out.pbf 2>/dev/null
+  rm ./out/out.geojson 2>/dev/null
+  echo "done."
+  echo
 
-# get latest osm data
-echo "downloading latest OSM data for Munich ..."
-if [ ! -f ./geo/oberbayern-latest.osm.pbf ]; then
-  curl --insecure https://download.geofabrik.de/europe/germany/bayern/oberbayern-latest.osm.pbf -o ./geo/oberbayern-latest.osm.pbf
+  # get latest osm data
+  echo "downloading latest OSM data for Munich ..."
+  if [ ! -f ./geo/oberbayern-latest.osm.pbf ]; then
+    curl --insecure https://download.geofabrik.de/europe/germany/bayern/oberbayern-latest.osm.pbf -o ./geo/oberbayern-latest.osm.pbf
+  else
+    echo "skipping since file already present."
+  fi
+  echo "done."
+  echo
+
+  # get latest annotations
+  echo "downloading latest munichways.com ratings ..."
+  if [ ! -f ./geo/munichways.json ]; then
+    curl --insecure "https://mw.mhcmuc.de/cgi-bin/qgis_mapserv.fcgi?SERVICE=WFS&VERSION=2.0.0&REQUEST=GetFeature&TYPENAME=Radvorrangnetz%2CRadvorrangnetz_gesamt%2Centlang_Hauptstrassen%2CGesamtnetz&outputFormat=application/json" -o geo/munichways.json
+    sed -i -e 's/"Name"/"name"/g' geo/munichways.json
+  else
+    echo "skipping since file already present."
+  fi
+  echo "done."
+  echo
+
+  # create out.pbf and out.geojson
+  echo "integrating ratings into OSM data ..."
+  python3 annotate_osm.py
+  echo "done."
+  echo
+
 else
-  echo "skipping since file already present."
-fi
-echo "done."
-echo
 
-# get latest annotations
-echo "downloading latest munichways.com ratings ..."
-if [ ! -f ./geo/munichways.json ]; then
-  curl --insecure "https://mw.mhcmuc.de/cgi-bin/qgis_mapserv.fcgi?SERVICE=WFS&VERSION=2.0.0&REQUEST=GetFeature&TYPENAME=Radvorrangnetz%2CRadvorrangnetz_gesamt%2Centlang_Hauptstrassen%2CGesamtnetz&outputFormat=application/json" -o geo/munichways.json
-  sed -i -e 's/"Name"/"name"/g' geo/munichways.json
-else
-  echo "skipping since file already present."
-fi
-echo "done."
-echo
+  echo "skipping annotation."
+  echo
 
-# create out.pbf and out.geojson
-echo "integrating ratings into OSM data ..."
-python3 annotate_osm.py
-echo "done."
-echo
+fi
 
 # prepare routing data
 echo "preparing data for routing ..."
